Исправляет баг с отрисовкой списка фильмов, переделывает реализацию закрытия попапа
